requirements.txt
Flask
ffmpeg-python
configparser
pymediainfo  # Optional, for rich metadata